requirements.txt
tensorflow>=2.10.0
streamlit>=1.25.0
numpy>=1.21.0
pillow>=9.0.0
scikit-learn>=1.0.0
matplotlib>=3.4.0
joblib>=1.1.0
